release
my project time limit is at an end and thanks to my groupmates to make this project possible. there will be bugs on the dark mode feature and needs some refinement but this will be improved on the next project. hope you could gained some insight on my code and tips to improved my writing. 

:)